Refactoring MainPageTemplate
* Moved MainPageTemplate to pages
* Created folder PagesByTemplates
* Moved PagesByMainPageTemplate to PagesByTemplates
