--- src/containers/index.tsx
@@ -1,1 +1,0 @@
-export { default as MainPageTemplate } from './MainPageTemplate'
