+++ src/pages/MainPageTemplate/MainPageTemplate.interfaces.tsx
