--- src/pages/PagesByMainPageTemplate/index.tsx
@@ -1,1 +1,0 @@
-export * from './PagesByMainPageTemplate'
