+++ src/pages/pagesByTemplates/byMainPageTemplate/byMainPageTemplate.interfaces.tsx
