+++ src/pages/pagesByTemplates/byMainPageTemplate/index.tsx
@@ -1,0 +1,1 @@
+export * from './byMainPageTemplate'
